+++ src/Action/Seo/SitemapAction.php
@@ -17,8 +17,8 @@ use Psr\Http\Server\RequestHandlerInterface;
 use Thepixeldeveloper\Sitemap\Drivers\XmlWriterDriver;
 use Thepixeldeveloper\Sitemap\Sitemap;
 use Thepixeldeveloper\Sitemap\SitemapIndex;
-use Zend\Diactoros\Response;
-use Zend\Diactoros\Stream;
+use Laminas\Diactoros\Response;
+use Laminas\Diactoros\Stream;
 
 class SitemapAction implements MiddlewareInterface
 {