+++ src/Package.php
@@ -9,55 +9,22 @@ declare(strict_types=1);
 
 namespace Ixocreate\Cms;
 
-use Ixocreate\Application\Configurator\ConfiguratorRegistryInterface;
 use Ixocreate\Application\Package\PackageInterface;
-use Ixocreate\Application\Service\ServiceRegistryInterface;
 use Ixocreate\Cms\Block\BlockBootstrapItem;
 use Ixocreate\Cms\PageType\PageTypeBootstrapItem;
-use Ixocreate\ServiceManager\ServiceManagerInterface;
 
 final class Package implements PackageInterface
 {
     /**
-     * @param ConfiguratorRegistryInterface $configuratorRegistry
+     * @return array
      */
-    public function configure(ConfiguratorRegistryInterface $configuratorRegistry): void
-    {
-    }
-
-    /**
-     * @param ServiceRegistryInterface $serviceRegistry
-     */
-    public function addServices(ServiceRegistryInterface $serviceRegistry): void
-    {
-    }
-
-    /**
-     * @return array|null
-     */
-    public function getBootstrapItems(): ?array
+    public function getBootstrapItems(): array
     {
         return [
             PageTypeBootstrapItem::class,
             BlockBootstrapItem::class,
             CmsBootstrapItem::class,
         ];
-    }
-
-    /**
-     * @return array|null
-     */
-    public function getConfigProvider(): ?array
-    {
-        return null;
-    }
-
-    /**
-     * @param ServiceManagerInterface $serviceManager
-     */
-    public function boot(ServiceManagerInterface $serviceManager): void
-    {
-        // TODO: Implement boot() method.
     }
 
     /**
@@ -69,18 +36,10 @@ final class Package implements PackageInterface
     }
 
     /**
-     * @return null|string
+     * @return array
      */
-    public function getConfigDirectory(): ?string
+    public function getDependencies(): array
     {
-        return null;
-    }
-
-    /**
-     * @return array|null
-     */
-    public function getDependencies(): ?array
-    {
-        return null;
+        return [];
     }
 }