+++ src/Seo/Sitemap/XmlSitemapProviderInterface.php
@@ -9,7 +9,7 @@ declare(strict_types=1);
 
 namespace Ixocreate\Cms\Seo\Sitemap;
 
-use Ixocreate\ServiceManager\NamedServiceInterface;
+use Ixocreate\Application\ServiceManager\NamedServiceInterface;
 
 interface XmlSitemapProviderInterface extends NamedServiceInterface
 {