+++ src/Seo/Sitemap/XmlSitemapProviderSubManager.php
@@ -9,8 +9,8 @@ declare(strict_types=1);
 
 namespace Ixocreate\Cms\Seo\Sitemap;
 
-use Ixocreate\ServiceManager\SubManager\SubManager;
+use Ixocreate\ServiceManager\SubManager\AbstractSubManager;
 
-final class XmlSitemapProviderSubManager extends SubManager
+final class XmlSitemapProviderSubManager extends AbstractSubManager
 {
 }