+++ src/Site/Tree/SearchSubManager.php
@@ -9,8 +9,8 @@ declare(strict_types=1);
 
 namespace Ixocreate\Cms\Site\Tree;
 
-use Ixocreate\ServiceManager\SubManager\SubManager;
+use Ixocreate\ServiceManager\SubManager\AbstractSubManager;
 
-class SearchSubManager extends SubManager
+class SearchSubManager extends AbstractSubManager
 {
 }