+++ tests/src/PackageTest.php
@@ -9,13 +9,10 @@ declare(strict_types=1);
 
 namespace Ixocreate\Test\Cms;
 
-use Ixocreate\Application\Configurator\ConfiguratorRegistryInterface;
-use Ixocreate\Application\Service\ServiceRegistryInterface;
 use Ixocreate\Cms\Block\BlockBootstrapItem;
 use Ixocreate\Cms\CmsBootstrapItem;
 use Ixocreate\Cms\Package;
 use Ixocreate\Cms\PageType\PageTypeBootstrapItem;
-use Ixocreate\ServiceManager\ServiceManagerInterface;
 use PHPUnit\Framework\TestCase;
 
 class PackageTest extends TestCase
@@ -25,23 +22,15 @@ class PackageTest extends TestCase
      */
     public function testPackage()
     {
-        $configuratorRegistry = $this->getMockBuilder(ConfiguratorRegistryInterface::class)->getMock();
-        $serviceRegistry = $this->getMockBuilder(ServiceRegistryInterface::class)->getMock();
-        $serviceManager = $this->getMockBuilder(ServiceManagerInterface::class)->getMock();
-
         $package = new Package();
-        $package->configure($configuratorRegistry);
-        $package->addServices($serviceRegistry);
-        $package->boot($serviceManager);
 
         $this->assertSame([
             PageTypeBootstrapItem::class,
             BlockBootstrapItem::class,
             CmsBootstrapItem::class,
         ], $package->getBootstrapItems());
-        $this->assertNull($package->getConfigDirectory());
-        $this->assertNull($package->getConfigProvider());
-        $this->assertNull($package->getDependencies());
+
         $this->assertDirectoryExists($package->getBootstrapDirectory());
+        $this->assertEmpty($package->getDependencies());
     }
 }